refactor(forms): split apart `model.ts`
In the forms package, the main abstractions for controls were previously
defined in a file `model.ts`, which contained several thousand lines of
code. This made changing the forms abstractions in any way a challenging
operation. Additionally, there were several import cycles involving this
file and other parts of the forms system.

In this commit, `model.ts` is broken out into a number of files. Several
strategies were used to effect this refactoring:

 * The interfaces of the main abstractions are interdependent, and are split
   out into a separate `model/api.ts` file.
 * Each interface is implemented by a private implementation class.
 * Constructors for each abstraction are defined with a separate constructor
   signature, and a `const` constructor declaration which is merged with
   the interface type.
 * The `_find` operation, which previously depended on `instanceof` checks
   with the actual constructor types to achieve different behavior for
   different types of controls, has been refactored to rely on an internal
   method on the `AbstractControl` hierarchy instead.

Together, these changes will make the forms package much more maintainable,
and pave the way for type-specific changes needed to strongly type the forms
abstractions.
