+++ packages/forms/src/directives/abstract_form_group_directive.ts
@@ -8,7 +8,7 @@
 
 import {Directive, OnDestroy, OnInit} from '@angular/core';
 
-import {FormGroup} from '../model';
+import {FormGroup} from '../model/api';
 
 import {ControlContainer} from './control_container';
 import {Form} from './form_interface';
