+++ packages/forms/src/directives/ng_form.ts
@@ -8,7 +8,8 @@
 
 import {AfterViewInit, Directive, EventEmitter, forwardRef, Inject, Input, Optional, Self} from '@angular/core';
 
-import {AbstractControl, FormControl, FormGroup, FormHooks} from '../model';
+import {AbstractControl, FormControl, FormHooks} from '../model/api';
+import {FormGroup} from '../model/form_group';
 import {composeAsyncValidators, composeValidators, NG_ASYNC_VALIDATORS, NG_VALIDATORS} from '../validators';
 
 import {ControlContainer} from './control_container';
