+++ packages/forms/src/directives/reactive_directives/form_control_name.ts
@@ -8,7 +8,7 @@
 
 import {Directive, EventEmitter, forwardRef, Host, Inject, Input, OnChanges, OnDestroy, Optional, Output, Self, SimpleChanges, SkipSelf} from '@angular/core';
 
-import {FormControl} from '../../model';
+import {FormControl} from '../../model/api';
 import {NG_ASYNC_VALIDATORS, NG_VALIDATORS} from '../../validators';
 import {AbstractFormGroupDirective} from '../abstract_form_group_directive';
 import {ControlContainer} from '../control_container';
