+++ packages/forms/src/forms.ts
@@ -42,7 +42,11 @@ export {NgSelectOption, SelectControlValueAccessor} from './directives/select_co
 export {SelectMultipleControlValueAccessor, ɵNgSelectMultipleOption} from './directives/select_multiple_control_value_accessor';
 export {AsyncValidator, AsyncValidatorFn, CheckboxRequiredValidator, EmailValidator, MaxLengthValidator, MaxValidator, MinLengthValidator, MinValidator, PatternValidator, RequiredValidator, ValidationErrors, Validator, ValidatorFn} from './directives/validators';
 export {FormBuilder} from './form_builder';
-export {AbstractControl, AbstractControlOptions, FormArray, FormControl, FormControlOptions, FormControlStatus, FormGroup, ɵFormControlCtor} from './model';
+export {AbstractControl} from './model/abstract_control';
+export {AbstractControlOptions, FormControlOptions, FormControlStatus} from './model/api';
+export {FormArray, ɵFormArrayCtor} from './model/form_array';
+export {FormControl, ɵFormControlCtor} from './model/form_control';
+export {FormGroup, ɵFormGroupCtor} from './model/form_group';
 export {NG_ASYNC_VALIDATORS, NG_VALIDATORS, Validators} from './validators';
 export {VERSION} from './version';
