+++ packages/forms/src/model/form_array.ts
@@ -1,0 +1,198 @@
+/**
+ * @license
+ * Copyright Google LLC All Rights Reserved.
+ *
+ * Use of this source code is governed by an MIT-style license that can be
+ * found in the LICENSE file at https://angular.io/license
+ */
+
+import {AsyncValidatorFn, ValidatorFn} from '../directives/validators';
+
+import {AbstractControl} from './abstract_control';
+import {AbstractControlOptions, FormArray as IFormArray} from './api';
+import {getRawValue} from './form_control';
+import {assertAllValuesPresent, assertControlPresent} from './form_group';
+import {pickAsyncValidators, pickValidators} from './util';
+
+const FormArrayImpl = class FormArray extends AbstractControl implements IFormArray {
+  constructor(
+      public controls: AbstractControl[],
+      validatorOrOpts?: ValidatorFn|ValidatorFn[]|AbstractControlOptions|null,
+      asyncValidator?: AsyncValidatorFn|AsyncValidatorFn[]|null) {
+    super(pickValidators(validatorOrOpts), pickAsyncValidators(asyncValidator, validatorOrOpts));
+    this._initObservables();
+    this._setUpdateStrategy(validatorOrOpts);
+    this._setUpControls();
+    this.updateValueAndValidity({
+      onlySelf: true,
+      // If `asyncValidator` is present, it will trigger control status change from `PENDING` to
+      // `VALID` or `INVALID`.
+      // The status should be broadcasted via the `statusChanges` observable, so we set `emitEvent`
+      // to `true` to allow that during the control creation process.
+      emitEvent: !!this.asyncValidator
+    });
+  }
+
+  at(index: number): AbstractControl {
+    return this.controls[this._adjustIndex(index)];
+  }
+
+  push(control: AbstractControl, options: {emitEvent?: boolean} = {}): void {
+    this.controls.push(control);
+    this._registerControl(control);
+    this.updateValueAndValidity({emitEvent: options.emitEvent});
+    this._onCollectionChange();
+  }
+
+  insert(index: number, control: AbstractControl, options: {emitEvent?: boolean} = {}): void {
+    this.controls.splice(index, 0, control);
+
+    this._registerControl(control);
+    this.updateValueAndValidity({emitEvent: options.emitEvent});
+  }
+
+  removeAt(index: number, options: {emitEvent?: boolean} = {}): void {
+    // Adjust the index, then clamp it at no less than 0 to prevent undesired underflows.
+    let adjustedIndex = this._adjustIndex(index);
+    if (adjustedIndex < 0) adjustedIndex = 0;
+
+    if (this.controls[adjustedIndex])
+      this.controls[adjustedIndex]._registerOnCollectionChange(() => {});
+    this.controls.splice(adjustedIndex, 1);
+    this.updateValueAndValidity({emitEvent: options.emitEvent});
+  }
+
+  setControl(index: number, control: AbstractControl, options: {emitEvent?: boolean} = {}): void {
+    // Adjust the index, then clamp it at no less than 0 to prevent undesired underflows.
+    let adjustedIndex = this._adjustIndex(index);
+    if (adjustedIndex < 0) adjustedIndex = 0;
+
+    if (this.controls[adjustedIndex])
+      this.controls[adjustedIndex]._registerOnCollectionChange(() => {});
+    this.controls.splice(adjustedIndex, 1);
+
+    if (control) {
+      this.controls.splice(adjustedIndex, 0, control);
+      this._registerControl(control);
+    }
+
+    this.updateValueAndValidity({emitEvent: options.emitEvent});
+    this._onCollectionChange();
+  }
+
+  get length(): number {
+    return this.controls.length;
+  }
+
+  override setValue(value: any[], options: {onlySelf?: boolean, emitEvent?: boolean} = {}): void {
+    assertAllValuesPresent(this, value);
+    value.forEach((newValue: any, index: number) => {
+      assertControlPresent(this, index);
+      this.at(index).setValue(newValue, {onlySelf: true, emitEvent: options.emitEvent});
+    });
+    this.updateValueAndValidity(options);
+  }
+
+  override patchValue(value: any[], options: {onlySelf?: boolean, emitEvent?: boolean} = {}): void {
+    // Even though the `value` argument type doesn't allow `null` and `undefined` values, the
+    // `patchValue` can be called recursively and inner data structures might have these values, so
+    // we just ignore such cases when a field containing FormArray instance receives `null` or
+    // `undefined` as a value.
+    if (value == null /* both `null` and `undefined` */) return;
+
+    value.forEach((newValue: any, index: number) => {
+      if (this.at(index)) {
+        this.at(index).patchValue(newValue, {onlySelf: true, emitEvent: options.emitEvent});
+      }
+    });
+    this.updateValueAndValidity(options);
+  }
+
+  override reset(value: any = [], options: {onlySelf?: boolean, emitEvent?: boolean} = {}): void {
+    this._forEachChild((control: AbstractControl, index: number) => {
+      control.reset(value[index], {onlySelf: true, emitEvent: options.emitEvent});
+    });
+    this._updatePristine(options);
+    this._updateTouched(options);
+    this.updateValueAndValidity(options);
+  }
+
+  getRawValue(): any[] {
+    return this.controls.map(control => getRawValue(control));
+  }
+
+  clear(options: {emitEvent?: boolean} = {}): void {
+    if (this.controls.length < 1) return;
+    this._forEachChild((control: AbstractControl) => control._registerOnCollectionChange(() => {}));
+    this.controls.splice(0);
+    this.updateValueAndValidity({emitEvent: options.emitEvent});
+  }
+
+  /**
+   * Adjusts a negative index by summing it with the length of the array. For very negative indices,
+   * the result may remain negative.
+   */
+  private _adjustIndex(index: number): number {
+    return index < 0 ? index + this.length : index;
+  }
+
+  /** @internal */
+  override _syncPendingControls(): boolean {
+    let subtreeUpdated = this.controls.reduce((updated: boolean, child: AbstractControl) => {
+      return child._syncPendingControls() ? true : updated;
+    }, false);
+    if (subtreeUpdated) this.updateValueAndValidity({onlySelf: true});
+    return subtreeUpdated;
+  }
+
+  /** @internal */
+  override _forEachChild(cb: (c: AbstractControl, index: number) => void): void {
+    this.controls.forEach((control: AbstractControl, index: number) => {
+      cb(control, index);
+    });
+  }
+
+  /** @internal */
+  override _updateValue(): void {
+    (this as {value: any}).value =
+        this.controls.filter((control) => control.enabled || this.disabled)
+            .map((control) => control.value);
+  }
+
+  /** @internal */
+  override _anyControls(condition: (c: AbstractControl) => boolean): boolean {
+    return this.controls.some((control: AbstractControl) => control.enabled && condition(control));
+  }
+
+  /** @internal */
+  _setUpControls(): void {
+    this._forEachChild((control: AbstractControl) => this._registerControl(control));
+  }
+
+  /** @internal */
+  override _allControlsDisabled(): boolean {
+    for (const control of this.controls) {
+      if (control.enabled) return false;
+    }
+    return this.controls.length > 0 || this.disabled;
+  }
+
+  /** @internal */
+  override _find(name: string|number): AbstractControl|null {
+    return this.at(name as number) || null;
+  }
+
+  private _registerControl(control: AbstractControl) {
+    control.setParent(this);
+    control._registerOnCollectionChange(this._onCollectionChange);
+  }
+};
+
+export interface ɵFormArrayCtor {
+  new(controls: AbstractControl[],
+      validatorOrOpts?: ValidatorFn|ValidatorFn[]|AbstractControlOptions|null,
+      asyncValidator?: AsyncValidatorFn|AsyncValidatorFn[]|null): IFormArray;
+}
+
+export const FormArray: ɵFormArrayCtor = FormArrayImpl;
+export type FormArray = IFormArray;
